Refactor: extract buildPrMarkdown into SummaryPrMarkdownBuilder and rename Summaries to Topics
